+++ docs/changelog.rst
@@ -2,11 +2,16 @@ Changelog
 =========
 
 
+1.2.1 (2014-06-13)
+------------------
+
+- Added ``strip_attr`` option to decide whether return the element attributes for parse result in :func:`lazyxml.loads` and :func:`lazyxml.load`.
+
 1.2 (2014-06-10)
 ----------------
 
 - `Sphinx <http://sphinx.pocoo.org/>`_ doc supported.
-- The ``fp`` parameter for :func:`lazyxml.dump` is no longer just as a string of filename, also supports `file` or `file-like` object.
+- The ``fp`` parameter for :func:`lazyxml.dump` is not just as a string of filename, also supports `file` or `file-like` object.
 - :class:`collections.defaultdict` and :class:`collections.OrderedDict` supported in :func:`lazyxml.dump` and :func:`lazyxml.dumps`.
 - :data:`types.GeneratorType` object and other iterable object that have `__iter__` and `next` attribute supported in :func:`lazyxml.dump` and :func:`lazyxml.dumps`.